Moved tests to separate directory
