+++ phpunit.xml
@@ -1,29 +1,19 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<!-- http://www.phpunit.de/manual/current/en/appendixes.configuration.html -->
-<phpunit
-        backupGlobals="false"
-        backupStaticAttributes="false"
-        bootstrap="src/lib/Tests/bootstrap.php"
-        convertErrorsToExceptions="true"
-        convertNoticesToExceptions="true"
-        convertWarningsToExceptions="true"
-        colors="false"
-        >
-    <php>
-        <ini name="error_reporting" value="-1" />
-    </php>
+<phpunit xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:noNamespaceSchemaLocation="http://schema.phpunit.de/8.5/phpunit.xsd"
+         bootstrap="vendor/autoload.php"
+         failOnWarning="true"
+         colors="true">
     <testsuites>
-        <testsuite name="EzSystemsEzRecommendationClientBundle tests suite">
-            <directory suffix="Test.php">src/lib/Tests</directory>
+        <testsuite name="Lib Tests">
+            <directory>tests/lib/</directory>
         </testsuite>
     </testsuites>
-
     <filter>
         <whitelist>
-            <directory>./</directory>
+            <directory>src</directory>
             <exclude>
-                <directory>vendor/</directory>
+                <directory>tests/bundle/</directory>
+                <directory>tests/lib/</directory>
             </exclude>
         </whitelist>
     </filter>
