+++ tests/lib/API/APIEndPointClassTest.php
@@ -1,0 +1,16 @@
+<?php
+
+/**
+ * @copyright Copyright (C) eZ Systems AS. All rights reserved.
+ * @license For full copyright and license information view LICENSE file distributed with this source code.
+ */
+declare(strict_types=1);
+
+namespace EzSystems\EzRecommendationClient\Tests\API;
+
+use EzSystems\EzRecommendationClient\API\AbstractAPI;
+
+class APIEndPointClassTest extends AbstractAPI
+{
+    const API_NAME = 'api-test';
+}
