+++ tests/lib/fixtures/directory/export_directory/.htpasswd
